move spec decode example script from example folder into vllm for reusablity
